requirements.txt
absl-py==1.4.0
astunparse==1.6.3
Bottleneck==1.3.5
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
Flask==2.2.3
flatbuffers==1.12
fonttools==4.25.0
gast==0.4.0
gevent==22.10.2
google-auth==2.17.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.1
grpcio==1.48.2
h5py==3.6.0
idna==3.4
importlib-metadata==6.5.0
importlib-resources==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
munkres==1.1.4
numexpr==2.8.4
numpy==1.22.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==1.5.3
Pillow==9.4.0
pip==23.0.1
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
seaborn==0.12.2
setuptools==66.0.0
six==1.16.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow-macos==2.9.2
termcolor==2.2.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
wheel==0.38.4
wrapt==1.15.0
zipp==3.15.0
zope.event==4.6
zope.interface==5.4.0
